make the assistant better, and default to just 2 parties
